--- src/dummyData/MessageData.tsx
@@ -1,58 +1,0 @@
-interface Messageinterface {
-  id: number;
-  img:string;
-  name: string;
-  sms: string;
-  time: string
-}
-export const Messages: Messageinterface[] = [
-  {
-    id: 1,
-    img:"../../public/images/message/teacher.png",
-    name: "Mr. Harby",
-    sms: "How are you?",
-    time:"4:15 pm",
-  },
-  {
-    id: 2,
-    img:"../../public/images/message/teacher.png",
-    name: "Mr. Harby",
-    sms: "How are you?",
-    time:"4:15 pm",
-  },
-  {
-    id: 3,
-    img:"../../public/images/message/teacher.png",
-    name: "Mr. Harby",
-    sms: "How are you?",
-    time:"4:15 pm",
-  },
-  {
-    id: 4,
-    img:"../../public/images/message/teacher.png",
-    name: "Mr. Harby",
-    sms: "How are you?",
-    time:"4:15 pm",
-  },
-  {
-    id: 5,
-    img:"../../public/images/message/teacher.png",
-    name: "Mr. Harby",
-    sms: "How are you?",
-    time:"4:15 pm",
-  },
-  {
-    id: 6,
-    img:"../../public/images/message/teacher.png",
-    name: "Mr. Harby",
-    sms: "How are you?",
-    time:"4:15 pm",
-  },
-  {
-    id: 7,
-    img:"../../public/images/message/teacher.png",
-    name: "Mr. Harby",
-    sms: "How are you?",
-    time:"4:15 pm",
-  },
-];
